Author Details Screen #3
- Fix deletion error. Links must be loaded before performing any operations
- Delete book's cover image and author's profile image conditionally

+++ lib/features/books/viewmodels/author_details_viewmodel.dart
@@ -62,8 +62,10 @@ class AuthorDetailsViewModel extends _$AuthorDetailsViewModel {
     final bool authorDeleted =
         await DatabaseRepository.instance.deleteAuthor(author);
     if (!authorDeleted) return false;
-    // Delete profile picture file only if the author was deleted succesfully.
-    await FilesRepository.instance.deleteFile(author.profilePicture);
+    if (author.profilePicture.isNotEmpty) {
+      // Delete profile picture file only if the author was deleted succesfully.
+      await FilesRepository.instance.deleteFile(author.profilePicture);
+    }
     return authorDeleted;
   }
